pyproject.toml
[project]
name = "sankhya-mcp"
version = "0.1.0"
description = "MCP server for Sankhya documentation — semantic search via sqlite-vec"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.2.0",
    "sqlite-vec>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "numpy>=1.26.0",
    "fastembed>=0.4.0",
]

[project.scripts]
sankhya-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
